+++ CHANGELOG.md
@@ -69,6 +69,7 @@ Howler.js now also has the concept of plugins. The core represents 100% compatib
 - `FIXED`: Stopping a stream is now possible, along with various other fixes.
 - `FIXED`: Audio on Chrome for Android no longer gets stuck after a period of inactivity.
 - `FIXED`: Fixed crash in iOS <9 webview.
+- `FIXED`: Fixed bug in iOS that can cause audio distortion when opening/closing browser.
 
 ## 1.1.25 (July 29, 2014)
 - `ADDED`: The `AudioContext` is now available on the global `Howler` object (thanks Matt DesLauriers).